separate ante implementation
